Revert rack-timeout deprecation
The rack-timeout integration is automatically required along with gem itself.
So there's no need to deprecate the original require name, which is
causing noise to users.
