+++ examples/rails-6.0/Gemfile
@@ -22,6 +22,8 @@ gem 'jbuilder', '~> 2.7'
 # Use Active Model has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
 
+gem 'rack-timeout'
+
 gem 'sentry-raven', path: "../../"
 
 # Use Active Storage variant
