+++ spec/raven/integrations/rack_timeout_spec.rb
@@ -1,15 +1,8 @@
 require "spec_helper"
 require "rack/timeout/base"
-require "raven/integrations/rack_timeout"
+require "raven/integrations/rack-timeout"
 
 RSpec.describe "Rack timeout" do
-  it "prints deprecation warning when requiring with dasherized filename" do
-    expect do
-      require "raven/integrations/rack-timeout"
-    end.to output(
-      "[Deprecation Warning] Dasherized filename \"raven/integrations/rack-timeout\" is deprecated and will be removed in 4.0; use \"raven/integrations/rack_timeout\" instead\n" # rubocop:disable Style/LineLength
-    ).to_stderr
-  end
   it "should have a raven_context method defined" do
     exc = Rack::Timeout::RequestTimeoutException.new("REQUEST_URI" => "This is a URI")
